update/设置自动打包脚本和路由
