+++ .github/workflows/main.yml
@@ -6,9 +6,9 @@ name: CI
 on:
   # Triggers the workflow on push or pull request events but only for the master branch
   push:
-    branches: [ octocat ]
+    branches: [octocat]
   pull_request:
-    branches: [ octocat ]
+    branches: [octocat]
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
@@ -34,3 +34,17 @@ jobs:
         run: |
           echo Add other actions to build,
           echo test, and deploy your project.
+      - name: Use Node.js ${{ matrix.node-version }}
+        uses: actions/setup-node@v1
+
+        with:
+
+        node-version: ${{ matrix.node-version }}
+
+      - name: Install dependencies
+
+        run: npm install
+
+      - name: build package
+
+        run: make default
