+++ src/services/routerService.ts
@@ -1,0 +1,6 @@
+export const RouterPaths = {
+  Home: '/',
+  Articles: '/Articles/:type/:name',
+  Marks: '/Marks',
+  PdfPreview: '/PdfPreview',
+}
